alterando implementacao da lista da tabela ascii

+++ Lista3/functions.h
@@ -19,6 +19,6 @@ vector<float> getNumbersFromInput(std::string, char);
 string decimalToBinary(unsigned);
 
 /*Definindo valores padrão para as variaveis*/
-void printAsciiTable(unsigned = 32, unsigned = 127, unsigned = 25);
+void printAsciiTable(short int = 32, unsigned short = 127, unsigned = 25);
 
 void printMenu();
